+++ CHANGELOG.md
@@ -2,14 +2,17 @@
 
 ## Unreleased
 * define tables. problems, answers, comments.
-* password hash
 * post logout
-* error messsages when register/login errors. flash?
+* flash for errors when register/login
 
 
 ## 0.2.0-SNAPSHOT
 ### Added
 * register
+* password hash
+* Logout
+### Changed
+* git unignore *.sql
 
 
 ## 0.1.1